refactor: 게시글 생성, 수정도 가능하게 변경 (FE-39)
